Update package.json
From PyPi:

---

 This email is notifying you of an upcoming deprecation that we have determined may affect you as a result of your recent upload to 'wordlift-client'.

In the future, PyPI will require all newly uploaded source distribution filenames to comply with PEP 625. Any source distributions already uploaded will remain in place as-is and do not need to be updated.

Specifically, your recent upload of 'wordlift-client-1.68.0.tar.gz' is incompatible with PEP 625 because it does not contain the normalized project name 'wordlift_client'.

In most cases, this can be resolved by upgrading the version of your build tooling to a later version that supports PEP 625 and produces compliant filenames.